Add legislator model and spec
In order to retrieve legislator data from the Sunlight OpenStates api,
this commit adds a legislator class. In addition, it adds a model
concern that defines some class methods that mirror activerecords find
and where. It also adds methods that can build an object based on the
response sent back. In additon, this commit sets up the guardfile to
correctly watch and run specs for a gem instead of a rails project. And
lastly, it sets up the faraday middleware we are using when requesting
data.